Queue pattern should have precence over autodiscover
