+++ mysql/datadog_checks/mysql/data/conf.yaml.example
@@ -377,12 +377,6 @@ instances:
     #
     # empty_default_hostname: false
 
-    ## @param disable_generic_tags - boolean - optional - default: false
-    ## Replaces generic tag such as `server` with `mysql_server` to avoid getting it mixed with
-    ## other integratons tags.
-    #
-    disable_generic_tags: true
-
 ## Log Section
 ##
 ## type - required - Type of log input source (tcp / udp / file / windows_event)
