+++ mysql/tests/tags.py
@@ -3,8 +3,8 @@
 # Licensed under a 3-clause BSD style license (see LICENSE)
 from . import common
 
-METRIC_TAGS = ['tag1:value1', 'tag2:value2']
-SC_TAGS = ['port:' + str(common.PORT), 'tag1:value1', 'tag2:value2']
-SC_TAGS_MIN = ['port:' + str(common.PORT)]
-SC_TAGS_REPLICA = ['port:' + str(common.SLAVE_PORT), 'tag1:value1', 'tag2:value2']
-SC_FAILURE_TAGS = ['port:unix_socket']
+METRIC_TAGS = ['tag1', 'tag2']
+SC_TAGS = ['server:' + common.HOST, 'port:' + str(common.PORT), 'tag1', 'tag2']
+SC_TAGS_MIN = ['server:' + common.HOST, 'port:' + str(common.PORT)]
+SC_TAGS_REPLICA = ['server:' + common.HOST, 'port:' + str(common.SLAVE_PORT), 'tag1', 'tag2']
+SC_FAILURE_TAGS = ['server:' + common.HOST, 'port:unix_socket']
