+++ mysql/tox.ini
@@ -47,3 +47,4 @@ setenv =
     maria10.2: MYSQL_VERSION=10.2
     # EOL 24 June 2025
     maria10.5: MYSQL_VERSION=10.5
+    DDEV_SKIP_GENERIC_TAGS_CHECK=true
